docs(kafka): provide example payloads
