+++ examples/snippets/kafka/advancedWorkingWithIdempotency.ts
@@ -6,7 +6,7 @@ import { DynamoDBPersistenceLayer } from '@aws-lambda-powertools/idempotency/dyn
 import { SchemaType, kafkaConsumer } from '@aws-lambda-powertools/kafka';
 import type { SchemaConfig } from '@aws-lambda-powertools/kafka/types';
 import { Logger } from '@aws-lambda-powertools/logger';
-import { User } from './samples/user.es6.generated.js'; // protobuf generated class
+import { com } from './samples/user.generated.js'; // protobuf generated class
 
 const logger = new Logger({ serviceName: 'kafka-consumer' });
 const persistenceStore = new DynamoDBPersistenceLayer({
@@ -16,14 +16,14 @@ const persistenceStore = new DynamoDBPersistenceLayer({
 const schemaConfig = {
   value: {
     type: SchemaType.PROTOBUF,
-    schema: User,
+    schema: com.example.User,
   },
 } satisfies SchemaConfig;
 
 const processRecord = makeIdempotent(
   async (user, topic, partition, offset) => {
     logger.info('processing user', {
-      userId: user.id,
+      user,
       meta: {
         topic,
         partition,
@@ -35,7 +35,7 @@ const processRecord = makeIdempotent(
 
     return {
       success: true,
-      userId: user.id,
+      user,
     };
   },
   {
