+++ examples/snippets/kafka/gettingStartedPrimitiveValues.ts
@@ -3,7 +3,7 @@ import { Logger } from '@aws-lambda-powertools/logger';
 
 const logger = new Logger({ serviceName: 'kafka-consumer' });
 
-export const handler = kafkaConsumer<string, { id: number; name: string }>(
+export const handler = kafkaConsumer<string, { name: string; age: number }>(
   async (event, _context) => {
     for (const record of event.records) {
       // Key is automatically decoded as UTF-8 string
@@ -14,8 +14,8 @@ export const handler = kafkaConsumer<string, { id: number; name: string }>(
       logger.info('received value', {
         key,
         product: {
-          id: value.id,
           name: value.name,
+          age: value.age,
         },
       });
     }
