+++ examples/snippets/kafka/samples/kafkaEventAvro.json
@@ -1,0 +1,19 @@
+{
+  "eventSource": "aws:kafka",
+  "eventSourceArn": "arn:aws:kafka:eu-west-3:123456789012:cluster/powertools-kafka-esm/f138df86-9253-4d2a-b682-19e132396d4f-s3",
+  "bootstrapServers": "boot-z3majaui.c3.kafka-serverless.eu-west-3.amazonaws.com:9098",
+  "records": {
+    "python-with-avro-doc-3": [
+      {
+        "topic": "python-with-avro-doc",
+        "partition": 3,
+        "offset": 0,
+        "timestamp": 1750547105187,
+        "timestampType": "CREATE_TIME",
+        "key": "MTIz",
+        "value": "AwBXT2qalUhN6oaj2CwEeaEWFFBvd2VydG9vbHMK",
+        "headers": []
+      }
+    ]
+  }
+}
