--- examples/snippets/kafka/samples/user.es6.generated.js
@@ -1,262 +1,0 @@
-/*eslint-disable block-scoped-var, id-length, no-control-regex, no-magic-numbers, no-prototype-builtins, no-redeclare, no-shadow, no-var, sort-vars*/
-import * as $protobuf from "protobufjs/minimal";
-
-// Common aliases
-const $Reader = $protobuf.Reader, $Writer = $protobuf.Writer, $util = $protobuf.util;
-
-// Exported root namespace
-const $root = $protobuf.roots["default"] || ($protobuf.roots["default"] = {});
-
-export const User = $root.User = (() => {
-
-    /**
-     * Properties of a User.
-     * @exports IUser
-     * @interface IUser
-     * @property {number|null} [id] User id
-     * @property {string|null} [name] User name
-     * @property {number|null} [price] User price
-     */
-
-    /**
-     * Constructs a new User.
-     * @exports User
-     * @classdesc Represents a User.
-     * @implements IUser
-     * @constructor
-     * @param {IUser=} [properties] Properties to set
-     */
-    function User(properties) {
-        if (properties)
-            for (let keys = Object.keys(properties), i = 0; i < keys.length; ++i)
-                if (properties[keys[i]] != null)
-                    this[keys[i]] = properties[keys[i]];
-    }
-
-    /**
-     * User id.
-     * @member {number} id
-     * @memberof User
-     * @instance
-     */
-    User.prototype.id = 0;
-
-    /**
-     * User name.
-     * @member {string} name
-     * @memberof User
-     * @instance
-     */
-    User.prototype.name = "";
-
-    /**
-     * User price.
-     * @member {number} price
-     * @memberof User
-     * @instance
-     */
-    User.prototype.price = 0;
-
-    /**
-     * Creates a new User instance using the specified properties.
-     * @function create
-     * @memberof User
-     * @static
-     * @param {IUser=} [properties] Properties to set
-     * @returns {User} User instance
-     */
-    User.create = function create(properties) {
-        return new User(properties);
-    };
-
-    /**
-     * Encodes the specified User message. Does not implicitly {@link User.verify|verify} messages.
-     * @function encode
-     * @memberof User
-     * @static
-     * @param {IUser} message User message or plain object to encode
-     * @param {$protobuf.Writer} [writer] Writer to encode to
-     * @returns {$protobuf.Writer} Writer
-     */
-    User.encode = function encode(message, writer) {
-        if (!writer)
-            writer = $Writer.create();
-        if (message.id != null && Object.hasOwnProperty.call(message, "id"))
-            writer.uint32(/* id 1, wireType 0 =*/8).int32(message.id);
-        if (message.name != null && Object.hasOwnProperty.call(message, "name"))
-            writer.uint32(/* id 2, wireType 2 =*/18).string(message.name);
-        if (message.price != null && Object.hasOwnProperty.call(message, "price"))
-            writer.uint32(/* id 3, wireType 1 =*/25).double(message.price);
-        return writer;
-    };
-
-    /**
-     * Encodes the specified User message, length delimited. Does not implicitly {@link User.verify|verify} messages.
-     * @function encodeDelimited
-     * @memberof User
-     * @static
-     * @param {IUser} message User message or plain object to encode
-     * @param {$protobuf.Writer} [writer] Writer to encode to
-     * @returns {$protobuf.Writer} Writer
-     */
-    User.encodeDelimited = function encodeDelimited(message, writer) {
-        return this.encode(message, writer).ldelim();
-    };
-
-    /**
-     * Decodes a User message from the specified reader or buffer.
-     * @function decode
-     * @memberof User
-     * @static
-     * @param {$protobuf.Reader|Uint8Array} reader Reader or buffer to decode from
-     * @param {number} [length] Message length if known beforehand
-     * @returns {User} User
-     * @throws {Error} If the payload is not a reader or valid buffer
-     * @throws {$protobuf.util.ProtocolError} If required fields are missing
-     */
-    User.decode = function decode(reader, length, error) {
-        if (!(reader instanceof $Reader))
-            reader = $Reader.create(reader);
-        let end = length === undefined ? reader.len : reader.pos + length, message = new $root.User();
-        while (reader.pos < end) {
-            let tag = reader.uint32();
-            if (tag === error)
-                break;
-            switch (tag >>> 3) {
-            case 1: {
-                    message.id = reader.int32();
-                    break;
-                }
-            case 2: {
-                    message.name = reader.string();
-                    break;
-                }
-            case 3: {
-                    message.price = reader.double();
-                    break;
-                }
-            default:
-                reader.skipType(tag & 7);
-                break;
-            }
-        }
-        return message;
-    };
-
-    /**
-     * Decodes a User message from the specified reader or buffer, length delimited.
-     * @function decodeDelimited
-     * @memberof User
-     * @static
-     * @param {$protobuf.Reader|Uint8Array} reader Reader or buffer to decode from
-     * @returns {User} User
-     * @throws {Error} If the payload is not a reader or valid buffer
-     * @throws {$protobuf.util.ProtocolError} If required fields are missing
-     */
-    User.decodeDelimited = function decodeDelimited(reader) {
-        if (!(reader instanceof $Reader))
-            reader = new $Reader(reader);
-        return this.decode(reader, reader.uint32());
-    };
-
-    /**
-     * Verifies a User message.
-     * @function verify
-     * @memberof User
-     * @static
-     * @param {Object.<string,*>} message Plain object to verify
-     * @returns {string|null} `null` if valid, otherwise the reason why it is not
-     */
-    User.verify = function verify(message) {
-        if (typeof message !== "object" || message === null)
-            return "object expected";
-        if (message.id != null && message.hasOwnProperty("id"))
-            if (!$util.isInteger(message.id))
-                return "id: integer expected";
-        if (message.name != null && message.hasOwnProperty("name"))
-            if (!$util.isString(message.name))
-                return "name: string expected";
-        if (message.price != null && message.hasOwnProperty("price"))
-            if (typeof message.price !== "number")
-                return "price: number expected";
-        return null;
-    };
-
-    /**
-     * Creates a User message from a plain object. Also converts values to their respective internal types.
-     * @function fromObject
-     * @memberof User
-     * @static
-     * @param {Object.<string,*>} object Plain object
-     * @returns {User} User
-     */
-    User.fromObject = function fromObject(object) {
-        if (object instanceof $root.User)
-            return object;
-        let message = new $root.User();
-        if (object.id != null)
-            message.id = object.id | 0;
-        if (object.name != null)
-            message.name = String(object.name);
-        if (object.price != null)
-            message.price = Number(object.price);
-        return message;
-    };
-
-    /**
-     * Creates a plain object from a User message. Also converts values to other types if specified.
-     * @function toObject
-     * @memberof User
-     * @static
-     * @param {User} message User
-     * @param {$protobuf.IConversionOptions} [options] Conversion options
-     * @returns {Object.<string,*>} Plain object
-     */
-    User.toObject = function toObject(message, options) {
-        if (!options)
-            options = {};
-        let object = {};
-        if (options.defaults) {
-            object.id = 0;
-            object.name = "";
-            object.price = 0;
-        }
-        if (message.id != null && message.hasOwnProperty("id"))
-            object.id = message.id;
-        if (message.name != null && message.hasOwnProperty("name"))
-            object.name = message.name;
-        if (message.price != null && message.hasOwnProperty("price"))
-            object.price = options.json && !isFinite(message.price) ? String(message.price) : message.price;
-        return object;
-    };
-
-    /**
-     * Converts this User to JSON.
-     * @function toJSON
-     * @memberof User
-     * @instance
-     * @returns {Object.<string,*>} JSON object
-     */
-    User.prototype.toJSON = function toJSON() {
-        return this.constructor.toObject(this, $protobuf.util.toJSONOptions);
-    };
-
-    /**
-     * Gets the default type url for User
-     * @function getTypeUrl
-     * @memberof User
-     * @static
-     * @param {string} [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
-     * @returns {string} The default type url
-     */
-    User.getTypeUrl = function getTypeUrl(typeUrlPrefix) {
-        if (typeUrlPrefix === undefined) {
-            typeUrlPrefix = "type.googleapis.com";
-        }
-        return typeUrlPrefix + "/User";
-    };
-
-    return User;
-})();
-
-export { $root as default };
