+++ examples/snippets/kafka/samples/user.generated.d.ts
@@ -1,0 +1,112 @@
+import * as $protobuf from "protobufjs";
+import Long = require("long");
+/** Namespace com. */
+export namespace com {
+
+    /** Namespace example. */
+    namespace example {
+
+        /** Properties of a User. */
+        interface IUser {
+
+            /** User name */
+            name?: (string|null);
+
+            /** User age */
+            age?: (number|null);
+        }
+
+        /** Represents a User. */
+        class User implements IUser {
+
+            /**
+             * Constructs a new User.
+             * @param [properties] Properties to set
+             */
+            constructor(properties?: com.example.IUser);
+
+            /** User name. */
+            public name: string;
+
+            /** User age. */
+            public age: number;
+
+            /**
+             * Creates a new User instance using the specified properties.
+             * @param [properties] Properties to set
+             * @returns User instance
+             */
+            public static create(properties?: com.example.IUser): com.example.User;
+
+            /**
+             * Encodes the specified User message. Does not implicitly {@link com.example.User.verify|verify} messages.
+             * @param message User message or plain object to encode
+             * @param [writer] Writer to encode to
+             * @returns Writer
+             */
+            public static encode(message: com.example.IUser, writer?: $protobuf.Writer): $protobuf.Writer;
+
+            /**
+             * Encodes the specified User message, length delimited. Does not implicitly {@link com.example.User.verify|verify} messages.
+             * @param message User message or plain object to encode
+             * @param [writer] Writer to encode to
+             * @returns Writer
+             */
+            public static encodeDelimited(message: com.example.IUser, writer?: $protobuf.Writer): $protobuf.Writer;
+
+            /**
+             * Decodes a User message from the specified reader or buffer.
+             * @param reader Reader or buffer to decode from
+             * @param [length] Message length if known beforehand
+             * @returns User
+             * @throws {Error} If the payload is not a reader or valid buffer
+             * @throws {$protobuf.util.ProtocolError} If required fields are missing
+             */
+            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): com.example.User;
+
+            /**
+             * Decodes a User message from the specified reader or buffer, length delimited.
+             * @param reader Reader or buffer to decode from
+             * @returns User
+             * @throws {Error} If the payload is not a reader or valid buffer
+             * @throws {$protobuf.util.ProtocolError} If required fields are missing
+             */
+            public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): com.example.User;
+
+            /**
+             * Verifies a User message.
+             * @param message Plain object to verify
+             * @returns `null` if valid, otherwise the reason why it is not
+             */
+            public static verify(message: { [k: string]: any }): (string|null);
+
+            /**
+             * Creates a User message from a plain object. Also converts values to their respective internal types.
+             * @param object Plain object
+             * @returns User
+             */
+            public static fromObject(object: { [k: string]: any }): com.example.User;
+
+            /**
+             * Creates a plain object from a User message. Also converts values to other types if specified.
+             * @param message User
+             * @param [options] Conversion options
+             * @returns Plain object
+             */
+            public static toObject(message: com.example.User, options?: $protobuf.IConversionOptions): { [k: string]: any };
+
+            /**
+             * Converts this User to JSON.
+             * @returns JSON object
+             */
+            public toJSON(): { [k: string]: any };
+
+            /**
+             * Gets the default type url for User
+             * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
+             * @returns The default type url
+             */
+            public static getTypeUrl(typeUrlPrefix?: string): string;
+        }
+    }
+}
