+++ examples/snippets/kafka/samples/user.json
@@ -1,0 +1,4 @@
+{
+  "name": "...",
+  "age": "..."
+}
