+++ examples/snippets/kafka/templates/gettingStartedWithMsk.yaml
@@ -4,8 +4,8 @@ Resources:
   KafkaConsumerFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: app.lambda_handler
-      Runtime: python3.13
+      Runtime: nodejs22.x
+      Handler: index.js
       Timeout: 30
       Events:
         MSKEvent:
