+++ packages/kafka/package.json
@@ -114,7 +114,6 @@
       ]
     }
   },
-  "private": true,
   "devDependencies": {
     "avro-js": "^1.12.0",
     "protobufjs": "^7.5.3",
